required token for deletion of blog

+++ part4/blog-list/controllers/blogs.js
@@ -1,38 +1,35 @@
-const jwt = require('jsonwebtoken');
+const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
-
-
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
 
-  if (!decodedToken.id) {
-    return response.status(401).json({error: 'token missing or invalid'})
+  if (decodedToken.id) {
+    const user = request.user;
+
+    const newBlog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id,
+    });
+
+    const savedBlog = await newBlog.save();
+    console.log(savedBlog);
+    user.blogs = user.blogs.concat(savedBlog.id);
+    await user.save();
+    response.status(201).json(savedBlog);
   }
-
-  const user = await User.findById(decodedToken.id);
-
-  const newBlog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: user._id,
-  });
-
-  const savedBlog = await newBlog.save();
-  user.blogs = user.blogs.concat(savedBlog._id);
-  await user.save();
-  response.status(201).json(savedBlog);
 });
 
 blogsRouter.put("/:id", async (request, response) => {
@@ -53,9 +50,19 @@ blogsRouter.put("/:id", async (request, response) => {
   response.status(200).json(updatedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(400).json({ error: "blog does not exist" });
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (decodedToken.id) {
+    if (blog.user.toString() === request.user.id) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    }
+  }
 });
 
 module.exports = blogsRouter;
